pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "logs"]

[project]
name = "filmmash-api"
version = "0.1.0"
dependencies = [
    "fastapi[standard]",
    "pydantic-settings",
    "gunicorn",
    "SQLAlchemy",
    "alembic",
    "asyncpg",
    "prometheus-client",
    "psutil",
    "pytest",
    "pytest-asyncio",
    "passlib[argon2]",
    "PyJWT"
]
requires-python = ">=3.12"
authors = [
    {name = "Pedro Garcia", email = "pedrofsgarcia.pro@gmail.com"}
]
maintainers = [
    {name = "Pedro Garcia", email = "pedrofsgarcia.pro@gmail.com"}
]
description = "Backend and API GATEWAY for the FilmMash App"
readme = "README.md"
license = "MIT"
license-files = []
keywords = []
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.12"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "bandit",
    "mypy",
    "pydantic",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "httpx",
    "types-psutil",
]

[project.urls]
Homepage = "https://github/pedro-fs-garcia/filmmash-api"
Documentation = "https://github/pedro-fs-garcia/filmmash-api/docs"
Repository = "https://github/pedro-fs-garcia/filmmash-api"


[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["venv", "__pycache__"]


[tool.ruff.lint]
select = [
    "E",  # Pycodestyle errors
    "W",  # Pycodestyle warnings
    "F",  # Pyflakes
    "I",  # Import order
    "N",  # Naming conventions
    "B",  # Bugbear (boas práticas e possíveis bugs)
    "C4", # Comprehensions
    "SIM",# Simplifiable code
    "UP", # Upgrades de código Python (melhorias automáticas)
]
ignore = [
    "F822",  # undefined name in __all__
    "F704",  # unused awaitable
    "F403",  # from module import *
]

[tool.ruff.lint.per-file-ignores]
"app/**/*.py" = ["F401", "F811", "F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true

[tool.mypy]
strict = true
python_version = "3.12"
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
plugins = ["pydantic.mypy"]
disable_error_code = ["misc"]

[[tool.mypy.overrides]]
module = ["jwt.*", "passlib.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.bandit]
exclude_dirs = ["tests", "venv"]
exclude_files = ["run.py"]
